Refact: WelcomeVC를 MVC로 리팩토링
