+++ iOS-Week3/iOS-Week3/ViewController/LoginViewController.swift
@@ -57,6 +57,7 @@ final class LoginViewController: UIViewController, UITextFieldDelegate {
     @objc func loginButtonDidTap() {
         pushToWelcomeVC()
     }
+    
     @objc func findIDButtonTapped() {
             print("아이디 찾기 버튼이 눌렸습니다.")
         }
